requirements.txt
yt-dlp
openai-whisper
langchain-text-splitters
langchain-community
langchain-core
chromadb
sentence-transformers
torch
streamlit